Added new core class BGG::Link to represent the bgg link type, id and value
I do not think this should be committed as written, it's meant to demonstrate the problem I have and one solution I've implemented.
It's not the only solution

The current default behavior of BGG::Game is to return the names of
various links (like artists, game_designers, mechanics and families), but to leave the associated id in the depths of the game_data hash.

The nature of link makes it ideal to have a single class represent numerous types of links, normalizing the type with local classes make in a snap to have Artist, Family, and Mechanic classes to represent certain types of links.